requirements.txt
myst-parser~=0.18.0
pydata-sphinx-theme==0.12.0rc1
ablog~=0.10.25
sphinx-design~=0.3.0
sphinx_subfigure~=0.2.4
sphinx-notfound-page~=0.8.3
sphinx-timeline~=0.2.1
